12TDDs Day 4: Monty Hall Problem
This was a tough one for me as there weren't any assertions provided. Reviewing
the wikipedia entry on the [Monty Hall
Problem](http://en.wikipedia.org/wiki/Monty_Hall_problem), it seemed apparent
that assertions could be
 * Winning probability approaches 1/3 if contestant does not switch doors
 * Winning probability approaches 2/3 if contestant switches doors

Rather than attempting to test against showing the results "in such a way as to
make it easy to compare the effects of each strategy", I simply tested against
the probability of winning based on the two contestant behavior scenarios.  I
think this satisfies the spirit of the challenge if not the letter of the
challenge.